pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "retail"
version = "0.1.0"
description = "Retail API for Square integration"
authors = [
    {name = "Alejandro", email = "alejandro@synvya.com"}
]
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "squareup>=35.0.0",
    "synvya-sdk>=0.1.0",
    "python-jose[cryptography]",
    "types-python-jose",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "alembic>=1.15.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["retail"]
exclude-package-data = {"*" = ["tests/*", "tests/**/*"]}

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "node-and-date" 

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["square.*", "synvya_sdk.*"]
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3